Move reducer, initialState and UserContext to store.js
