+++ src/modules/profile/containers/ProfileContainer.jsx
@@ -1,7 +1,11 @@
-import React from 'react';
+import React, { useContext } from 'react';
+import { UserContext } from '../../../store';
+import Profile from '../components/Profile';
 
-const ProfileContainer = () => {
-	return <div></div>;
-};
+function ProfileContainer() {
+	const context = useContext(UserContext);
+
+	return <Profile user={context.state.user} />;
+}
 
 export default ProfileContainer;
