+++ src/pages/DashboardPage.jsx
@@ -5,7 +5,7 @@ import Messages from '../modules/dashboard/components/Messages';
 import Navbar from '../modules/dashboard/components/navbar/Navbar';
 import Profile from '../modules/dashboard/components/Profile';
 import Sidebar from '../modules/dashboard/components/sidebar/Sidebar';
-import ProfileContainer from '../modules/profile/components/ProfileContainer';
+import ProfileContainer from '../modules/profile/containers/ProfileContainer';
 
 const DashboardPage = () => {
 	return (
